Refactor seed file to have image, solution, chanllenge data in separate files
