+++ server/db/seed/image.js
@@ -1,0 +1,38 @@
+const images = [
+  {
+    type: 'challenge',
+    url: 'challenge-1-red-circle.png',
+    connector: 'challenge-1',
+    width: 100,
+    height: 100,
+  },
+  {
+    type: 'challenge',
+    url: 'challenge-2-blue-square.png',
+    connector: 'challenge-2',
+    width: 100,
+    height: 100,
+  },
+  {
+    type: 'challenge',
+    url: 'challenge-3-yellow-rectangle.png',
+    connector: 'challenge-3',
+    width: 200,
+    height: 100,
+  },
+  {
+    type: 'challenge',
+    url: 'challenge-4-orange-rectangle.png',
+    connector: 'challenge-4',
+    width: 100,
+    height: 200,
+  },
+  { type: 'userchallenge', url: 'userchallenge-1.png', connector: 'userchallenge-1' },
+  { type: 'userchallenge', url: 'userchallenge-2.png', connector: 'userchallenge-2' },
+  { type: 'userchallenge', url: 'userchallenge-3.png', connector: 'userchallenge-3' },
+  { type: 'userchallenge', url: 'userchallenge-4.png', connector: 'userchallenge-4' },
+];
+
+module.exports = {
+  images,
+};
