+++ server/db/seed/user.js
@@ -1,10 +1,27 @@
 const usersSeed = [
-  { email: 'grant@geezemail.com', password: '1234' },
-  { email: 'kristy@geezemail.com', password: '1234' },
-  { email: 'haoyu@geezemail.com', password: '1234' },
-  { email: 'thee@geezemail.com', password: '1234' },
-  { email: 'a@a.com', password: 'a' },
-  { email: 'b@b.com', password: 'b' },
+  {
+    firstName: 'Grant',
+    lastName: 'H',
+    email: 'grant@geezemail.com',
+    password: '1234',
+    type: 'user',
+  },
+  {
+    firstName: 'Kristy',
+    lastName: 'C',
+    email: 'kristy@geezemail.com',
+    password: '1234',
+    type: 'user',
+  },
+  {
+    firstName: 'Haoyu',
+    lastName: 'Y',
+    email: 'haoyu@geezemail.com',
+    password: '1234',
+    type: 'user',
+  },
+  { firstName: 'Theo', lastName: 'M', email: 'thee@geezemail.com', password: '1234', type: 'user' },
+  { firstName: 'Admin', lastName: 'Admin', email: 'a@a.com', password: 'a', type: 'admin' },
 ];
 
 module.exports = {
